requirements.txt
Flask==2.3.2
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.3
Flask-JWT-Extended==4.5.2
Flask-Migrate==4.0.5
Flask-WTF==1.1.1
Flask-CORS==4.0.0
SQLAlchemy==2.0.15
psycopg2==2.9.6
Pillow==9.5.0
requests==2.31.0
python-dotenv==1.0.0
gunicorn==21.2.0
email-validator==2.3.0